Remove hardcoded part for saving results

+++ app/backend/app/db/models.py
@@ -30,7 +30,7 @@ submissions = Table(
 users = Table(
     "users",
     metadata,
-    Column("id", BigInteger, primary_key=True),
+    Column("id", Integer, primary_key=True),
     Column("name", String(50)),
     Column("email", String(50)),
     Column("hashed_password", String(250)),
